Implemented apis with documentation hints

+++ src/main/java/za/co/investorWithdrawal/service/repository/UserInfoRepository.java
@@ -4,6 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import za.co.investorWithdrawal.service.repository.entity.UserInfoEntity;
 
+
 @Repository
-public interface UserInfoRepository extends JpaRepository<UserInfoEntity, Integer> {
+public interface UserInfoRepository extends JpaRepository<UserInfoEntity, Long> {
 }
